fix: 项目标准化
