+++ src/game/Ht.java
@@ -12,54 +12,53 @@ import java.util.Vector;
  * 为了让panel不停的重绘子弹，需要将ht 实现Run able，当作一个线程使用
  */
 public class Ht extends JPanel implements KeyListener, Runnable {
-    //定义自己的坦克
-    ziji ho = null;
-    //定义敌人坦克，放入Vector
+    // 定义自己的坦克
+    PlayerThank playerThank = null;
+    // 定义敌人坦克，放入Vector
     Vector<EnemyTank> enemyTanks = new Vector<>();
-    //定义一个Vector，用于存放炸弹图片
-    //当子弹击中坦克时，就加入一个bumb对象到boms
+    // 定义一个Vector，用于存放炸弹图片, 当子弹击中坦克时，就加入一个bumb对象到boms
     Vector<Bomb> bombs = new Vector<>();
-    //定义三张图片，用于显示爆炸效果
+    // 定义三张图片，用于显示爆炸效果
     Image image1 = null;
     Image image2 = null;
     Image image3 = null;
     int enemyTankSize = 3;
 
     public Ht() {
-        ho = new ziji(100, 100);//初始化自己的坦克
-        ho.setSpeed(2);
+        // 初始化自己的坦克
+        playerThank = new PlayerThank(100, 100);
+        playerThank.setSpeed(2);
         EnemyTank enemyTank;
-        //初始化敌人的坦克
+        // 初始化敌人的坦克
         for (int i = 0; i < enemyTankSize; i++) {
-            //初始化敌人的坦克并且指定方向
+            // 初始化敌人的坦克并且指定方向
             enemyTank = new EnemyTank((100 * (i + 1)), 0);
-            //指定方向
+            // 指定方向
             enemyTank.setDirect(2);
 
+            enemyTank.setSpeed(2);
+
             new Thread(enemyTank).start();
-            //给enemy Tank加入一颗子弹
+            // 给enemy Tank加入一颗子弹
             Shot shot = new Shot(enemyTank.getX() + 20, enemyTank.getY() + 60, enemyTank.getDirect());
-            //加入enemytank的Vector 成员
+            // 加入enemytank的Vector 成员
             enemyTank.shots.add(shot);
-            //启动shot对象
+            // 启动shot对象
             new Thread(shot).start();
-            //加入数组
+            // 加入数组
             enemyTanks.add(enemyTank);
         }
 
         //初始化图片对象
         /*
-        * 这段代码是Java中获取图像资源的方法。
-        * 具体来说，Toolkit类是Java中用于访问本地窗口系统的工具类，
-        * getDefaultToolkit()是获取默认Toolkit实例的静态方法。
-        * 然后，通过Panel类的getResource()方法获取资源路径，
-        * 再通过getImage()方法获取图像资源。
-        * "/bomb_1.gif"是资源路径，表示在程序运行时，
-        * 从项目的根目录开始寻找名为"bomb_1.gif"的图像文件。
-总之，这段代码的作用是获取指定路径下的图像资源，并将其存储在Image对象中，以便在程序中使用。
-        *
-        *
-        * */
+         * 这段代码是Java中获取图像资源的方法。
+         * 具体来说，Toolkit类是Java中用于访问本地窗口系统的工具类，
+         * 然后，通过Panel类的getResource()方法获取资源路径，
+         * 再通过getImage()方法获取图像资源。
+         * "/bomb_1.gif"是资源路径，表示在程序运行时，
+         * 从项目的根目录开始寻找名为"bomb_1.gif"的图像文件。
+         * 总之，这段代码的作用是获取指定路径下的图像资源，并将其存储在Image对象中，以便在程序中使用。
+         */
         image1 = Toolkit.getDefaultToolkit().getImage(Panel.class.getResource("/bomb_1.gif"));
         image2 = Toolkit.getDefaultToolkit().getImage(Panel.class.getResource("/bomb_2.gif"));
         image3 = Toolkit.getDefaultToolkit().getImage(Panel.class.getResource("/bomb_3.gif"));
@@ -70,16 +69,16 @@ public class Ht extends JPanel implements KeyListener, Runnable {
         super.paint(g);
         g.fillRect(0, 0, 1000, 750);//填充矩形，默认黑色
         //画出坦克-封装方法
-        drawTank(ho.getX(), ho.getY(), g, ho.getDirect(), 1);
+        drawTank(playerThank.getX(), playerThank.getY(), g, playerThank.getDirect(), 1);
         //画出ho射击子弹
-        if (ho.shot != null && ho.shot.isLive == true) {
-//            g.fill3DRect(ho.shot.x,ho.shot.y,1,1,false);
-            g.draw3DRect(ho.shot.x, ho.shot.y, 3, 3, false);
+        if (playerThank.shot != null && playerThank.shot.isLive == true) {
+//            g.fill3DRect(playerThank.shot.x,playerThank.shot.y,1,1,false);
+            g.draw3DRect(playerThank.shot.x, playerThank.shot.y, 3, 3, false);
         }
-        //如果bombs集合中有对象就画出爆炸效果
+        // 如果bombs集合中有对象就画出爆炸效果
         for (int i = 0; i < bombs.size(); i++) {
             Bomb bomb = bombs.get(i);
-            //根据当前bomb对象的life值画出对应的图片
+            // 根据当前bomb对象的life值画出对应的图片
             if (bomb.life > 6) {
                 g.drawImage(image1, bomb.x, bomb.y, 60, 60, this);
             } else if (bomb.life > 3) {
@@ -87,26 +86,28 @@ public class Ht extends JPanel implements KeyListener, Runnable {
             } else {
                 g.drawImage(image3, bomb.x, bomb.y, 60, 60, this);
             }
-            //让炸弹生命值减少
+            // 让炸弹生命值减少
             bomb.lifeDown();
-            //如果bomb为0 就从bombs集合中删除
+            // 如果bomb为0 就从bombs集合中删除
             if (bomb.life == 0) {
                 bombs.remove(bomb);
             }
         }
-        //画出敌人的坦克，遍历vector
+        // 画出敌人的坦克，遍历vector
         for (int i = 0; i < enemyTanks.size(); i++) {
-            //取出坦克
+            // 取出坦克
             EnemyTank enemyTank = enemyTanks.get(i);
-            //然后调用敌人坦克的方法
-            if (enemyTank.isLeve) {//判断敌人坦克是否为false如果是false就不绘制
+            // 然后调用敌人坦克的方法
+            // 判断敌人坦克是否为false如果是false就不绘制
+            if (enemyTank.isLeve) {
                 drawTank(enemyTank.getX(), enemyTank.getY(), g, enemyTank.getDirect(), 0);
-                //画出enemyTank 所有子弹
+                // 画出enemyTank 所有子弹
                 for (int j = 0; j < enemyTank.shots.size(); j++) {
-                    //取出当前enemytan的k子弹
+                    // 取出当前enemytan的k子弹
                     Shot shot = enemyTank.shots.get(j);
-                    //绘制
-                    if (shot.isLive) { // isLive == true才绘制子弹
+                    // 绘制
+                    // isLive == true才绘制子弹
+                    if (shot.isLive) {
                         g.draw3DRect(shot.x, shot.y, 3, 3, false);
                     } else {
                         //从Vector移出子弹
@@ -116,6 +117,7 @@ public class Ht extends JPanel implements KeyListener, Runnable {
             }
         }
     }
+
     //编写方法，画出坦克
 
     /**
@@ -126,19 +128,22 @@ public class Ht extends JPanel implements KeyListener, Runnable {
      * @param type   坦克的类型
      */
     public void drawTank(int x, int y, Graphics g, int direct, int type) {
-        //判断坦克类型
-        //根据不同的坦克类型，设置不同的颜色
+        // 判断坦克类型
+        // 根据不同的坦克类型，设置不同的颜色
         switch (type) {
+            // 表示敌人的坦克
             case 0:
-                //表示敌人的坦克
                 g.setColor(Color.cyan);
                 break;
-            case 1://表示自己的坦克
+            // 表示自己的坦克
+            case 1:
                 g.setColor(Color.yellow);
                 break;
+            default:
+                throw new IllegalStateException("Unexpected value: " + type);
         }
-        //根据坦克方向，来绘制坦克
-        //direct 表示方向(0：表示向上，1表示向右，2表示向下，3表示向左)
+        // 根据坦克方向，来绘制坦克
+        // direct 表示方向(0：表示向上，1表示向右，2表示向下，3表示向左)
         switch (direct) {
             case 0://表示向上
                 g.fill3DRect(x, y, 10, 60, false);//画出坦克左边的轮子
@@ -174,35 +179,47 @@ public class Ht extends JPanel implements KeyListener, Runnable {
         }
     }
 
-    //编写方法判断我方子弹是否击中地方坦克
+    /**
+     * 编写方法判断我方子弹是否击中地方坦克
+     *
+     * @param s
+     * @param enemyTank
+     */
     public void hitTank(Shot s, EnemyTank enemyTank) {
-        //判断s是否击中坦克
-        switch (enemyTank.getDirect()) { //enemyTank.getDirect()传入敌人坦克的上下左右进行判断
-            case 0://坦克向上
-            case 2://坦克向下
+        // 判断s是否击中坦克
+        // enemyTank.getDirect()传入敌人坦克的上下左右进行判断
+        switch (enemyTank.getDirect()) {
+            // 坦克向上
+            case 0:
+                // 坦克向下
+            case 2:
                 if (s.x > enemyTank.getX() && s.x < enemyTank.getX() + 40 && s.y > enemyTank.getY() && s.y < enemyTank.getY() + 60) {
-                    //如果满足以上条件就销毁子弹线程和坦克
+                    // 如果满足以上条件就销毁子弹线程和坦克
                     s.isLive = false;
                     enemyTank.isLeve = false;
-                    //当我的子弹击中敌人坦克后，将enemytan从kvector拿掉
+                    // 当我的子弹击中敌人坦克后，将enemytan从kvector拿掉
                     enemyTanks.remove(enemyTank);
-                    //创建bomb对象，加入到bobms集合
+                    // 创建bomb对象，加入到bobms集合
                     Bomb bomb = new Bomb(enemyTank.getX(), enemyTank.getY());
-                    //创建好一个对象后加入到boms里面去
+                    // 创建好一个对象后加入到boms里面去
                     bombs.add(bomb);
                 }
                 break;
-            case 1://向右
-            case 3://向左
+            // 向右
+            case 1:
+            // 向左
+            case 3:
                 if (s.x > enemyTank.getX() && s.x < enemyTank.getX() + 60 && s.y > enemyTank.getY() && s.y < enemyTank.getY() + 40) {
                     s.isLive = false;
                     enemyTank.isLeve = false;
                     enemyTanks.remove(enemyTank);
                     Bomb bomb = new Bomb(enemyTank.getX(), enemyTank.getY());
-                    //创建好一个对象后加入到boms里面去
+                    // 创建好一个对象后加入到boms里面去
                     bombs.add(bomb);
                 }
-
+                break;
+            default:
+                throw new IllegalStateException("Unexpected value: " + enemyTank.getDirect());
         }
     }
 
@@ -211,27 +228,31 @@ public class Ht extends JPanel implements KeyListener, Runnable {
 
     }
 
-    //处理键wdsa 按下的情况
+    /**
+     * 处理键wdsa 按下的情况
+     *
+     * @param e
+     */
     @Override
     public void keyPressed(KeyEvent e) {
         if (e.getKeyCode() == KeyEvent.VK_W) {//按下w键
             //改变坦克的方向
-            ho.setDirect(0);
+            playerThank.setDirect(0);
             //修改坦克的坐标
-            ho.moveUP();
+            playerThank.moveUP();
         } else if (e.getKeyCode() == KeyEvent.VK_D) {//D键
-            ho.setDirect(1);
-            ho.moveRight();
+            playerThank.setDirect(1);
+            playerThank.moveRight();
         } else if (e.getKeyCode() == KeyEvent.VK_S) {//S键
-            ho.setDirect(2);
-            ho.moveDown();
+            playerThank.setDirect(2);
+            playerThank.moveDown();
         } else if (e.getKeyCode() == KeyEvent.VK_A) {//A键
-            ho.setDirect(3);
-            ho.moveLeft();
+            playerThank.setDirect(3);
+            playerThank.moveLeft();
         }
         //如果用户按下的是j，就发射
         if (e.getKeyCode() == KeyEvent.VK_J) {
-            ho.Shotziji();
+            playerThank.Shotziji();
         }
         //让面板重绘
         this.repaint();
@@ -252,12 +273,12 @@ public class Ht extends JPanel implements KeyListener, Runnable {
             }
             //判断是否击中了敌人坦克因为这个每隔一百秒就会执行一次所以把上面的扔下来
             //这里要判断一下这个是否创建了
-            if (ho.shot != null && ho.shot.isLive) {//判断自己的子弹是否还活着
+            if (playerThank.shot != null && playerThank.shot.isLive) {//判断自己的子弹是否还活着
                 //因为不知道你击中了敌人的哪个坦克所以说用遍历
                 for (int i = 0; i < enemyTanks.size(); i++) {
                     //取出敌人坦克
                     EnemyTank enemyTank = enemyTanks.get(i);
-                    hitTank(ho.shot, enemyTank);
+                    hitTank(playerThank.shot, enemyTank);
                 }
             }
             this.repaint();
