--- src/game/tk.java
@@ -1,58 +1,0 @@
-package game;
-//坦克
-public class tk {
-    private int x;//坦克的横坐标
-    private int y;//坦克的纵坐标
-    private int direct;//坦克方向0，上1，右2，下3左
-    private int speed;//控制坦克速度
-
-    public int getSpeed() {
-        return speed;
-    }
-
-    public void setSpeed(int speed) {
-        this.speed = speed;
-    }
-
-    //因为私有的原因无法调用所以说要写以恶搞公共的改变坦克坐标的方法
-    public void moveUP(){
-        y -= speed;
-    }
-    public void moveRight(){
-        x += speed;
-    }
-    public void moveDown(){
-        y += speed;
-    }
-    public void moveLeft(){
-        x -= speed;
-    }
-    public int getDirect() {
-        return direct;
-    }
-
-    public void setDirect(int direct) {
-        this.direct = direct;
-    }
-
-    public tk(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-
-    public int getX() {
-        return x;
-    }
-
-    public void setX(int x) {
-        this.x = x;
-    }
-
-    public int getY() {
-        return y;
-    }
-
-    public void setY(int y) {
-        this.y = y;
-    }
-}
